% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxnorm_in.R
\docType{data}
\name{rxnorm_in}
\alias{rxnorm_in}
\title{Drug ingredients in RXNORM}
\format{
A data frame with 6443 rows and 3 variables:
\describe{
\item{IN_RXCUI}{RXCUI for the drug ingredient or multi-ingredient (MIN)}
\item{IN_TTY}{Type of ingredient (IN, MIN)}
\item{IN_STR}{Name of drug ingredient}
}
}
\usage{
rxnorm_in
}
\description{
A dataset containing RXNORM drug ingredients (IN) and
combination drugs as multi-ingredients (MIN) available in the
RXNORM current prescribable content requiring no licenses for use.  The
subset excludes suppressed and obsolete drugs, drugs only prescribed
outside the US, and drugs only for veterinary use.
}
\seealso{
\link{rxnorm_in_map}

\url{https://www.nlm.nih.gov/research/umls/rxnorm/docs/rxnormfiles.html}
}
\keyword{datasets}
