% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxnorm_in.R
\docType{data}
\name{rxnorm_in_map}
\alias{rxnorm_in_map}
\title{Mapping of drug ingredients to component drugs and other identifiers}
\format{
A data frame with 555023 rows and 6 variables:
\describe{
\item{IN_RXCUI}{RXCUI for the drug ingredient or multi-ingredient (MIN)}
\item{IN_TTY}{Type of ingredient (IN, MIN)}
\item{IN_STR}{Name of drug ingredient}
\item{drug_code}{Drug code}
\item{drug_name}{Name of component drug in RXNORM.  NA for NDC and HCPCS}
\item{drug_type}{Type of drug component/code, including: \itemize{
\item{SCD: Semantic clinical drug = Ingredient + Strength + Dose
form}
\item{SBD: Semantic branded drug =  Ingredient + Strength + Dose
form+Brand name}
\item{SCDC: Semantic Clinical Drug Component = Ingredient + Strength}
\item{SCDF: Semantic Clinical Drug Form = Ingredient + Dose form}
\item{SCDG: Semantic Clinical Drug Group = Ingredient + Dose form
group}
\item{DP: Drug Pack = Ingredient + Strength}
\item{NDC: National Drug Codes, packaging code.  Includes both RXNORM
normalized code (without dashes) and non-normalized codes}
\item{HCPCS: Healthcare Common Procedure Coding System J codes}
}
}
}
}
\usage{
rxnorm_in_map
}
\description{
A dataset containing mappings of component drugs and codes to RXNORM drug
ingredients (IN) extracted from the RXNORM relationships (RXNREL) and
attributes (RXNSAT) tables.
Mappings are available for RXNORM component types (SCD, SBD, SCDC, SCDF,
SCDG, DP), to National Drug Codes (NDC) and Healthcare Common Procedure
Coding System J codes (HCPCS).  These mappings do not have license
restrictions.
}
\keyword{datasets}
