% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxnorm_mesh_indications.R
\docType{data}
\name{rxnorm_mesh_indications}
\alias{rxnorm_mesh_indications}
\title{Mapping of drug ingredients to MeSH indications}
\format{
A data frame with 7950 rows and 5 variables:
\describe{
\item{IN_RXCUI}{RXCUI for the drug ingredient or multi-ingredient (MIN)}
\item{IN_STR}{Name of drug ingredient}
\item{RELA}{Type of indication, either may_treat, may_prevent or
may_diagnose}
\item{MESH_ID}{MeSH disease identifier}
\item{MESH_STR}{MeSH disease name}
}
}
\usage{
rxnorm_mesh_indications
}
\description{
A dataset containing mappings of RXNORM drug ingredients to Medical Subject
Headings (MeSH) indications extracted from RxMix.
}
\keyword{datasets}
